Change invincibleUntilChapter to double

+++ app/src/main/java/hu/dj/aradventure/armodel/Enemy.kt
@@ -7,7 +7,7 @@ open class Enemy: ArModel() {
     open var damagePoint: Int = 0
     open var reward: Item? = null
     open var loopDeathAnimation: Boolean = true
-    open var invincibleUntilChapter: Float = 0.0F
+    open var invincibleUntilChapter: Double = 0.0
 
     fun damage(damagePoint: Int) {
         this.health -= damagePoint
